Modularise googleMyBusiness
Also moves `ui/google-my-business` under the main state key
for `googleMyBusiness`.

Fixes #42436
