+++ client/my-sites/google-my-business/stats/chart.js
@@ -22,10 +22,10 @@ import PieChartLegend from 'components/pie-chart/legend';
 import PieChartLegendPlaceholder from 'components/pie-chart/legend-placeholder';
 import PieChartPlaceholder from 'components/pie-chart/placeholder';
 import SectionHeader from 'components/section-header';
-import { changeGoogleMyBusinessStatsInterval } from 'state/ui/google-my-business/actions';
+import { changeGoogleMyBusinessStatsInterval } from 'state/google-my-business/ui/actions';
 import { enhanceWithSiteType, recordTracksEvent } from 'state/analytics/actions';
 import { getSelectedSiteId } from 'state/ui/selectors';
-import { getStatsInterval } from 'state/ui/google-my-business/selectors';
+import { getStatsInterval } from 'state/google-my-business/ui/selectors';
 import { requestGoogleMyBusinessStats } from 'state/google-my-business/actions';
 import { withEnhancers } from 'state/utils';
