+++ client/state/ui/reducer.js
@@ -14,7 +14,6 @@ import checkout from './checkout/reducer';
 import comments from './comments/reducer';
 import dropZone from './drop-zone/reducer';
 import editor from './editor/reducer';
-import googleMyBusiness from './google-my-business/reducer';
 import guidedTour from './guided-tours/reducer';
 import editorDeprecationDialog from './editor-deprecation-dialog/reducer';
 import gutenbergOptInDialog from './gutenberg-opt-in-dialog/reducer';
@@ -105,7 +104,6 @@ const reducer = combineReducers( {
 	dropZone,
 	editor,
 	editorDeprecationDialog,
-	googleMyBusiness,
 	guidedTour,
 	gutenbergOptInDialog,
 	hasSidebar,
